+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/services/local/BankInterface.java
@@ -92,7 +92,7 @@ public class BankInterface {
 	}
 
 	@Atomic(mode = TxMode.WRITE)
-	public static void deposit(String iban, double amount) {
+	public static void deposit(String iban, long amount) {
 		Account account = getAccountByIban(iban);
 		if (account == null) {
 			throw new BankException();
@@ -102,7 +102,7 @@ public class BankInterface {
 	}
 
 	@Atomic(mode = TxMode.WRITE)
-	public static void withdraw(String iban, double amount) {
+	public static void withdraw(String iban, long amount) {
 		Account account = getAccountByIban(iban);
 		if (account == null) {
 			throw new BankException();
